+++ src/core/components.ts
@@ -2,12 +2,14 @@
 
 import * as ng from 'angular';
 import * as breadcrumbModule from '../components/breadcrumb/breadcrumbDirective';
-import * as calloutModule from '../components/callout/calloutDirective';
 import * as buttonModule from '../components/button/buttonDirective';
+import * as calloutModule from '../components/callout/calloutDirective';
 import * as choicefieldModule from '../components/choicefield/choicefieldDirective';
-import * as contextualMenuModule from '../components/contextualmenu/contextualMenu';
 import * as commandBarModule from '../components/commandbar/commandBarDirective';
+import * as contentModule from '../components/content/contentDirective';
+import * as contextualMenuModule from '../components/contextualmenu/contextualMenu';
 import * as datepickerModule from '../components/datepicker/datepickerDirective';
+import * as dialogModule from '../components/dialog/dialogDirective';
 import * as dropdownModule from '../components/dropdown/dropdownDirective';
 import * as iconModule from '../components/icon/iconDirective';
 import * as labelModule from '../components/label/labelDirective';
@@ -20,7 +22,6 @@ import * as spinnerModule from '../components/spinner/spinnerDirective';
 import * as tableModule from '../components/table/tableDirective';
 import * as textFieldModule from '../components/textfield/textFieldDirective';
 import * as toggleModule from '../components/toggle/toggleDirective';
-import * as contentModule from '../components/content/contentDirective';
 
 /**
  * @ngdoc module
@@ -33,12 +34,14 @@ import * as contentModule from '../components/content/contentDirective';
  */
 export var module: ng.IModule = ng.module('officeuifabric.components', [
   breadcrumbModule.module.name,
-  calloutModule.module.name,
   buttonModule.module.name,
+  calloutModule.module.name,
   choicefieldModule.module.name,
-  contextualMenuModule.module.name,
   commandBarModule.module.name,
+  contentModule.module.name,
+  contextualMenuModule.module.name,
   datepickerModule.module.name,
+  dialogModule.module.name,
   dropdownModule.module.name,
   iconModule.module.name,
   labelModule.module.name,
@@ -50,6 +53,5 @@ export var module: ng.IModule = ng.module('officeuifabric.components', [
   spinnerModule.module.name,
   tableModule.module.name,
   textFieldModule.module.name,
-  toggleModule.module.name,
-  contentModule.module.name
+  toggleModule.module.name
 ]);